# docs/api.yaml
...
openapi: 3.1.0
info:
  title: Harness API
  version: 0.2.0
  description: |
    The HTTP surface of the Harness process.

    Authentication is checked before existence: an unauthenticated request to an
    authenticated route never learns whether the path exists. Every JSON failure
    shares one envelope (`ApiError`) carrying a human sentence, a stable machine
    code, and whether a retry may help.
servers:
  - url: http://127.0.0.1:8080
    description: Default local bind address.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Master token from `HARNESS_AUTH_TOKEN`, or a short-lived session token
        issued by `POST /auth/session`.
  schemas:
    ApiError:
      type: object
      description: |
        The single failure envelope for every JSON route. `error` is a complete
        human sentence intended to be shown verbatim; it is the field the browser
        UI reads. `code` is the stable token clients branch on. `retryable` says
        whether repeating the identical request may succeed.
      required: [error, code, retryable]
      additionalProperties: false
      properties:
        error:
          type: string
          description: Human-readable sentence, safe to display verbatim.
        code:
          type: string
          description: |
            Stable machine-readable code. Derived from the HTTP status by
            `error_code` in `src/api/error.rs`; this enum must match that table
            exactly.
          enum:
            - invalid_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - gone
            - payload_too_large
            - unsupported_media_type
            - unprocessable_body
            - rate_limited
            - internal_error
            - not_implemented
            - upstream_failed
            - unavailable
            - upstream_timeout
            - error
        retryable:
          type: boolean
          description: |
            True only for congestion and upstream faults (429, 502, 503, 504).
            A 500 is deliberately false: the server cannot promise the failed
            operation had no effect. Clients that retry anyway must rely on the
            request identifier for idempotency.
...
  responses:
    Error:
      description: Request failed; body is the shared error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
security:
  - bearerAuth: []
paths:
...
  /sessions:
    get:
      operationId: sessions
      summary: List conversations.
      responses:
        '200':
          description: Sessions.
          content:
            application/json: {}
        '4XX':
          $ref: '#/components/responses/Error'
        '5XX':
          $ref: '#/components/responses/Error'
...
  /memory/ingest:
    post:
      operationId: ingestMemory
      summary: Ingest source material for memory extraction.
      description: Accepts a larger body than other routes (2 MiB).
      responses:
        '200':
          description: Ingest receipt.
          content:
            application/json: {}
        '202':
          description: Ingest accepted for background work.
          content:
            application/json: {}
        '4XX':
          $ref: '#/components/responses/Error'
        '5XX':
          $ref: '#/components/responses/Error'
...
  /changes:
    get:
      operationId: changes
      summary: List recorded file changes.
      responses:
        '200':
          description: File changes.
          content:
            application/json: {}
        '4XX':
          $ref: '#/components/responses/Error'
        '5XX':
          $ref: '#/components/responses/Error'
...